+++ src/p_inter.c
@@ -1444,7 +1444,7 @@ void P_DamageMobj(mobj_t *target, mobj_t *inflicter, mobj_t *source, int damage,
 
         tplayer->damagecount = damagecount;
 
-        if (gp_vibrate && vibrate)
+        if (gp_vibrate_damage && vibrate)
         {
             XInputVibration(30000 + (100 - MIN(tplayer->health, 100)) / 100 * 30000);
             damagevibrationtics += BETWEEN(12, damage, 100);