+++ src/p_mobj.c
@@ -1176,7 +1176,7 @@ void P_SpawnPuff(fixed_t x, fixed_t y, fixed_t z, angle_t angle)
         P_SetMobjState(th, S_PUFF3);
 
         // [BH] vibrate XInput gamepads
-        if (gp_vibrate && vibrate)
+        if (gp_vibrate_damage && vibrate)
         {
             int motorspeed = weaponinfo[wp_fist].motorspeed;